+++ test/development/basic/asset-prefix/fixture/next.config.js
