+++ test/development/basic/asset-prefix/fixture/pages/index.js
