+++ test/development/basic/barrel-optimization/fixture/app/client-boundary/page.js
