+++ test/development/basic/barrel-optimization/fixture/app/client/page.js
