+++ test/development/basic/barrel-optimization/fixture/app/dedupe/page.js
