+++ test/development/basic/barrel-optimization/fixture/app/layout.js
