+++ test/development/basic/barrel-optimization/fixture/app/page.js
