+++ test/development/basic/barrel-optimization/fixture/app/recursive/shortcut/page.js
