+++ test/development/basic/barrel-optimization/fixture/components/slow-component.js
