+++ test/development/basic/barrel-optimization/fixture/node_modules_bak/my-client-lib/index.js
