+++ test/development/basic/barrel-optimization/fixture/pages/recursive-barrel.js
