+++ test/development/basic/define-class-fields/fixture/pages/index.tsx
