+++ test/development/basic/define-class-fields/fixture/tsconfig.json
