+++ test/development/basic/emotion-swc/fixture/jsconfig.json
