+++ test/development/basic/emotion-swc/fixture/pages/_app.js
