+++ test/development/basic/emotion-swc/fixture/pages/index.js
