+++ test/development/basic/emotion-swc/fixture/shared/styles.js
