+++ test/development/basic/theme-ui/fixture/jsconfig.json
